+++ conf/urls.py
@@ -14,6 +14,7 @@ from django.views.generic import RedirectView
 import core.views
 import export_readiness.views
 from activitystream.views import ActivityStreamView
+from groups.views import GroupInfoModalView
 
 api_router = WagtailAPIRouter('api')
 api_router.register_endpoint('pages', core.views.PagesOptionalDraftAPIEndpoint)
@@ -100,6 +101,11 @@ urlpatterns = [
         ),
         login_required(csrf_exempt(core.views.PreloadPageView.as_view())),
         name='preload-add-page',
+    ),
+    url(
+        r'^admin/group-info/$',
+        login_required(GroupInfoModalView.as_view()),
+        name='group-info',
     ),
 
     # Prevent users from changing their email address