+++ users/views.py
@@ -1,14 +1,17 @@
+from django.conf import settings
 from django.contrib.auth import get_user_model, update_session_auth_hash
 from django.shortcuts import redirect
-from django.urls import reverse
+from django.urls import reverse, reverse_lazy
+from django.utils import timezone
 from django.utils.translation import ugettext as _
 from django.views import generic
-
 from wagtail.admin import messages
+from wagtail.admin.views.generic import EditView
 from wagtail.core import hooks
 from wagtail.users.utils import user_can_delete_user
 from wagtail.users.views.users import add_user_perm, change_user_perm
 
+from users.models import UserProfile
 from . import forms, mixins
 
 User = get_user_model()
@@ -85,6 +88,32 @@ class EditUserView(
     form_invalid_message = 'The user could not be saved due to errors.'
     form_valid_message = 'User ''{0}'' updated.'
 
+    def dispatch(self, request, *args, **kwargs):
+        self.object = self.get_object()
+        profile = self.object.userprofile
+
+        self.is_approval = profile.assignment_status in (
+            UserProfile.STATUS_CREATED,
+            UserProfile.STATUS_AWAITING_APPROVAL,
+        )
+        if self.is_approval and request.method == 'GET':
+            # Warn the current user that this is an 'approval'
+            message_text = (
+                "This user is awaiting approval and will be automatically "
+                "notified via email if added to a group that grants access "
+                "to the Wagtail CMS."
+            )
+            if profile.self_assigned_group:
+                # Preselect the user-selcted group in the form
+                self.initial['groups'] = [profile.self_assigned_group.id]
+                # Also mention this to the current user
+                message_text += (
+                    " They requested to be added to the '{}' group, so this "
+                    "has been preselected for you under the 'Roles' tab."
+                ).format(profile.self_assigned_group.name)
+            messages.warning(request, message_text)
+        return super().dispatch(request, *args, **kwargs)
+
     def get_form_kwargs(self):
         editing_self = self.request.user == self.object
         kwargs = super().get_form_kwargs()
@@ -93,10 +122,16 @@ class EditUserView(
 
     def form_valid(self, form):
         self.object = user = form.save()
+
         if user == self.request.user and 'password1' in form.changed_data:
             # User is changing their own password;
             # need to update their session hash
             update_session_auth_hash(self.request, user)
+
+        if self.is_approval and user.has_perm('wagtailadmin.access_admin'):
+            self.mark_user_as_approved()
+            self.notify_user_of_approval()
+
         self.handle_success_message()
         return redirect('wagtailusers_users:index')
 
@@ -105,3 +140,53 @@ class EditUserView(
         can_delete = user_can_delete_user(self.request.user, self.object)
         ctx.update(can_delete=can_delete)
         return ctx
+
+    def mark_user_as_approved(self):
+        profile = self.object.userprofile
+        profile.assignment_status = UserProfile.STATUS_APPROVED
+        profile.approved_by = self.request.user
+        profile.approved_at = timezone.now()
+        profile.save()
+
+    def notify_user_of_approval(self):
+        # TODO: connect to gov.notify
+        pass  # pragma: no cover
+
+
+class SSORequestAccessView(EditView):
+    model = UserProfile
+    form_class = forms.SSORequestAccessForm
+    permission_required = None
+    error_message = 'There was a problem with your submission'
+    template_name = "sso/request_access.html"
+    success_url = reverse_lazy('wagtailusers_users:sso_request_access_success')
+
+    def dispatch(self, request):
+        self.get_object()
+        if settings.USERS_REQUEST_ACCESS_PREVENT_RESUBMISSION:
+            # Toggling behaviour here to aid local testing
+            if self.object.assignment_status == UserProfile.STATUS_APPROVED:
+                return redirect('wagtailadmin_home')
+            if self.object.assignment_status == UserProfile.STATUS_AWAITING_APPROVAL: # noqa
+                return redirect(self.success_url)
+        return super().dispatch(request)
+
+    def get_object(self):
+        if hasattr(self, 'object'):
+            return self.object
+        self.object = self.request.user.userprofile
+        return self.object
+
+    def save_instance(self):
+        self.object = super().save_instance()
+        self.object.assignment_status = UserProfile.STATUS_AWAITING_APPROVAL
+        self.object.save()
+        self.notify_team_leader_of_pending_approval()
+        return self.object
+
+    def get_success_url(self):
+        return self.success_url
+
+    def notify_team_leader_of_pending_approval(self):
+        # TODO: connect to gov.notify
+        pass  # pragma: no cover